requirements/base.txt
Pillow==11.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.7.0  # https://github.com/evansd/whitenoise
redis==5.1.1  # https://github.com/redis/redis-py

# Django
# ------------------------------------------------------------------------------
django  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-filter
django-environ
requests
psycopg2
django-cleanup

# Django REST Framework
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt
django-cors-headers
django-silk
channels
daphne
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
django-environ
pre-commit
